requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2

# Google Cloud
google-cloud-aiplatform==1.70.0
google-cloud-storage==2.18.0
google-cloud-sql-connector==1.13.0
google-cloud-firestore==2.19.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Google APIs
google-api-python-client==2.147.0
google-api-core==2.19.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
sqlmodel==0.0.22
alembic==1.13.2

# Redis for caching/session
redis==5.1.1
aioredis==2.0.1

# Background tasks
celery==5.4.0

# HTTP client
httpx==0.27.2
aiofiles==24.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Monitoring & Logging
prometheus-client==0.21.0
structlog==24.4.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.8.0
isort==5.13.2
mypy==1.11.2

# Additional utilities
python-dotenv==1.0.1
Pillow==10.4.0
pandas==2.2.3
numpy==2.1.1
